ci(pytest): Update pytest for CvDeviceTools
