+++ tests/unit/data.py
@@ -1,0 +1,20 @@
+CVP_DEVICES = [
+    {
+        "fqdn": "CV-ANSIBLE-EOS01",
+        "systemMacAddress": "50:8d:00:e3:78:aa",
+        "parentContainerName": "ANSIBLE",
+        "configlets": [
+                "CV-EOS-ANSIBLE01"
+        ],
+        "imageBundle": []
+    },
+    {
+        "serialNumber": "79AEA53101E7340AEC9AA4819D5E1F5B",
+        "systemMacAddress": "50:8d:00:e3:78:aa",
+        "parentContainerName": "ANSIBLE",
+        "configlets": [
+                "CV-EOS-ANSIBLE01"
+        ],
+        "imageBundle": []
+    }
+]
